Fix: Map progression generation fixed
wasn’t correctly generating in the interval between segments

+++ Assets/Scripts/General/Utility/Interval.cs
@@ -4,5 +4,5 @@ using System.Collections;
 [System.Serializable]
 public class Interval
 {
-	public int Start, End, Delta;
+	public float Start, End, Delta;
 }
